Fixed types in convertLinopParams

+++ pyQCD/core/kernel/include/utils.hpp
@@ -106,7 +106,7 @@ namespace pyQCD
 #ifdef USE_CUDA
   void convertLinopParams(int* cudaIntParams, const vector<int> intParams,
 			  float* cudaFloatParams,
-			  const vector<float> floatParams,
+			  const vector<double> floatParams,
 			  Complex* cudaComplexParams,
 			  const vector<complex<double> > complexParams);
